+++ src/key_storage/macos_key_storage.rs
@@ -1,13 +1,11 @@
-#![cfg(target_os = "macos")]
-
 use std::borrow::Cow;
 
 use enostr::Keypair;
 use serde::{Deserialize, Serialize};
 
-use crate::file_key_storage::BasicFileStorage;
-use crate::key_storage::{KeyStorage, KeyStorageResponse};
-use crate::security_framework_key_storage::SecurityFrameworkKeyStorage;
+use super::file_key_storage::BasicFileStorage;
+use super::key_storage_impl::{KeyStorage, KeyStorageResponse};
+use super::security_framework_key_storage::SecurityFrameworkKeyStorage;
 use crate::settings::StorageSettings;
 
 pub struct MacOSKeyStorage<'a> {
@@ -25,7 +23,7 @@ impl<'a> KeyStorage for MacOSKeyStorage<'a> {
         match &self.settings.macos_key_storage_type {
             MacOSKeyStorageType::BasicFileStorage => BasicFileStorage::new().get_keys(),
             MacOSKeyStorageType::SecurityFramework(service_name) => {
-                SecurityFrameworkKeyStorage::new(&service_name).get_keys()
+                SecurityFrameworkKeyStorage::new(service_name).get_keys()
             }
         }
     }
@@ -34,7 +32,7 @@ impl<'a> KeyStorage for MacOSKeyStorage<'a> {
         match &self.settings.macos_key_storage_type {
             MacOSKeyStorageType::BasicFileStorage => BasicFileStorage::new().add_key(key),
             MacOSKeyStorageType::SecurityFramework(service_name) => {
-                SecurityFrameworkKeyStorage::new(&service_name).add_key(key)
+                SecurityFrameworkKeyStorage::new(service_name).add_key(key)
             }
         }
     }
@@ -43,7 +41,7 @@ impl<'a> KeyStorage for MacOSKeyStorage<'a> {
         match &self.settings.macos_key_storage_type {
             MacOSKeyStorageType::BasicFileStorage => BasicFileStorage::new().remove_key(key),
             MacOSKeyStorageType::SecurityFramework(service_name) => {
-                SecurityFrameworkKeyStorage::new(&service_name).remove_key(key)
+                SecurityFrameworkKeyStorage::new(service_name).remove_key(key)
             }
         }
     }
