+++ src/lib.rs
@@ -18,9 +18,7 @@ mod frame_history;
 mod images;
 mod imgcache;
 mod key_parsing;
-mod key_storage;
 pub mod login_manager;
-mod macos_key_storage;
 mod multi_subscriber;
 mod nav;
 mod note;
@@ -44,11 +42,10 @@ mod view_state;
 #[cfg(test)]
 #[macro_use]
 mod test_utils;
-mod linux_key_storage;
 
-mod file_key_storage;
-mod security_framework_key_storage;
 mod settings;
+
+mod key_storage;
 
 pub use app::Damus;
 pub use error::Error;
